+++ app/models/legal_services/journey/check_suitability.rb
@@ -5,7 +5,6 @@ module LegalServices
     SUITABILITY_OPTIONS = %w[yes no].freeze
 
     attribute :service_suitable
-    attribute :lot
     validates :service_suitable, inclusion: SUITABILITY_OPTIONS
   end
 end