+++ app/models/legal_services/journey/choose_services.rb
@@ -2,12 +2,16 @@ module LegalServices
   class Journey::ChooseServices
     include Steppable
 
-    attribute :lot
-    attribute :services, Array
-    validates :services, presence: true
+    attribute :lot_id
+    attribute :service_ids, Array
+    validates :service_ids, presence: true
 
-    def selected_lot
-      service_name::Lot.find_by(number: lot)
+    def lot_services
+      Service.where(lot_id:).order(:name)
+    end
+
+    def lot
+      Lot.find(lot_id)
     end
   end
 end