+++ app/models/management_consultancy/rm6309/journey/choose_lot.rb
@@ -1,18 +1,6 @@
 module ManagementConsultancy
   module RM6309
-    class Journey::ChooseLot
-      include Steppable
-
-      attribute :lot
-      validates :lot, inclusion: { in: Lot.all_numbers }
-
-      def self.mcf4_lots
-        Lot.where(framework: 'MCF4').sort_by { |lot| lot.number[5..].to_i }
-      end
-
-      def next_step_class
-        Journey::ChooseServices
-      end
+    class Journey::ChooseLot < ManagementConsultancy::Journey::ChooseLot
     end
   end
 end