+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Small change...
+  In all seriousness this is something I’ve wanted to do for a while.
+  I have made generic framework and supplier models that can be used by all the frameworks in this service.
+  This should make managing and adding new services and frameworks easier going forward ([PR 1996](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/1996))
+
 ## [5.1.0] - 2025-06-30
 
 ### Added